+++ changelog.md
@@ -1,5 +1,4 @@
 # Changelog
 
 **./readme.md**
-* Moved all request method endpoints to separate lines.
-* Even if they are the same URL.
+* Removed horizontal lines